pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Ali-Fattahian <ali.fattahian1378@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.7"
django = "3.2"
asgiref = "3.6.0"
certifi = "2023.5.7"
charset-normalizer = "3.1.0"
cloudinary = "1.32.0"
django-cloudinary-storage = "0.3.0"
django-cors-headers = "4.0.0"
django-filter = "23.2"
djangorestframework = "3.14.0"
djangorestframework-simplejwt = "5.2.2"
idna = "3.4"
Pillow = "9.5.0"
psycopg2 = "2.9.6"
PyJWT = "2.7.0"
python-dotenv = "0.20.0"
pytz = "2023.3"
requests = "2.30.0"
six = "1.16.0"
sqlparse = "0.4.4"
urllib3 = "1.26.15"
whitenoise = {extras = ["brotli"], version = "^6.4.0"}
gunicorn = "^20.1.0"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
